wait before hover
When hovering over a particular DOM element to display an icon and then clicking on that icon, the icon will be hidden and the click may fail if you don’t wait for loading images before hovering. This problem is now resolved automatically.

The test fails because chrome caches an image, so I made it so that it generates a random path every time.

I tried https://stackoverflow.com/questions/29988227/disable-cache-in-selenium-chrome-driver but I couldn't turn off the cache.